Moved dependency and install logic out of main module manager
- All auto-install/upgrade and dependency logic now in AdvancedManager
- Unit tests updated
- More verbose parameter name for Zend\Module\Manager::getMergedConfig()
